+++ src/data/export.ts
@@ -1,5 +1,5 @@
 import {format} from 'date-fns/fp'
-import * as fs from 'fs'
+import * as fs from 'node:fs'
 import * as ExcelJS from 'exceljs'
 
 export default class Export {
@@ -8,16 +8,17 @@ export default class Export {
   public static async write(
     name: string,
     data: {
-      matched: string;
-      case_name: string;
-      assertion_name: string;
-      expected: string;
-      actual: string;
+      matched: string
+      case_name: string
+      assertion_name: string
+      expected: string
+      actual: string
     }[],
   ): Promise<string> {
     if (!fs.existsSync('reports')) {
       fs.mkdirSync('reports')
     }
+
     const workbook = new ExcelJS.Workbook()
     workbook.creator = 'wate - Web API Testing tool'
     workbook.title = name
@@ -57,7 +58,7 @@ export default class Export {
         fgColor: {argb: '008810'},
       }
 
-      columns.forEach(name => {
+      for (const name of columns) {
         const currentCell = row.getCell(name)
         currentCell.fill =
           row.getCell('matched').value === '✓' ? fillSuccess : fillError
@@ -65,10 +66,14 @@ export default class Export {
           size: 12,
           color: {argb: 'FFFFFFFF'},
         }
-      })
+      }
     })
     // Align columns
-    worksheet.columns.every(col => (col.alignment = {vertical: 'top'}))
+    worksheet.columns.every(col => {
+      col.alignment = {vertical: 'top'}
+
+      return col
+    })
     // Wrap text for expected and asserted
     const expectedCol = worksheet.getColumn('expected')
     expectedCol.alignment = {wrapText: true, horizontal: 'right'}