+++ src/suite.ts
@@ -1,15 +1,31 @@
 import {KeyValue as ParamKeyValue} from './param'
 import {KeyValue as CaptureKeyValue} from './capture'
-import {KeyValue as AssertionKeyValue} from './assertion'
+import {AssertionKeyValue} from './assertion'
 import Request from './request'
 
 export interface SuiteDefinition {
   name: string
+  matrix?: Matrix[]
   cases: CaseDefinition[]
+}
+
+interface Matrix {
+  name: string
+  caseName: string
+  params: ParamKeyValue
+  captures: CaptureKeyValue
+  assertions: AssertionKeyValue
+}
+
+export interface KeyValueBag {
+  params: ParamKeyValue
+  captures: CaptureKeyValue
+  assertions: AssertionKeyValue
 }
 
 export interface CaseDefinition {
   name: string
+  matrix?: boolean
   requests: RequestDefinition[]
 }