+++ src/suite/loader.ts
@@ -1,16 +1,16 @@
 import * as path from 'node:path'
 import * as fs from 'node:fs'
 import * as yaml from 'js-yaml'
-import {Suite, Case, SuiteDefinition, RequestDefinition} from '../suite'
+import {JSON_SCHEMA} from 'js-yaml'
+import {Suite, Case, SuiteDefinition, RequestDefinition, CaseDefinition, KeyValueBag} from '../suite'
 import Request from '../request'
 import SchemaValidator, {ValidationSchema} from '../validator/schema'
 import Context from '../context'
 import RequestBuilder from '../request/builder'
-import Param from '../param'
 import DataHelper from '../data/helper'
-import CaptureDefinition from '../capture'
-import AssertionDefinition from '../assertion'
-import {JSON_SCHEMA} from 'js-yaml'
+import Param, {KeyValue as ParamKeyValue} from '../param'
+import CaptureDefinition, {KeyValue as CaptureKeyValue} from '../capture'
+import AssertionDefinition, {KeyValue as AssertionKeyValue} from '../assertion'
 
 export default class SuiteLoader {
   static schema: ValidationSchema = {
@@ -18,6 +18,13 @@ export default class SuiteLoader {
     properties: {
       name: {type: 'string'},
       description: {type: 'string'},
+      matrix: {
+        type: 'array',
+        items: {
+          $ref: '#/definitions/matrix',
+        },
+        minItems: 0,
+      },
       cases: {
         type: 'array',
         items: {
@@ -29,10 +36,33 @@ export default class SuiteLoader {
     required: ['name'],
     additionalProperties: false,
     definitions: {
+      matrix: {
+        type: 'object',
+        properties: {
+          name: {type: 'string'},
+          caseName: {type: 'string'},
+          description: {type: 'string'},
+          params: {
+            type: 'object',
+            additionalProperties: true,
+          },
+          captures: {
+            type: 'object',
+            additionalProperties: true,
+          },
+          assertions: {
+            type: 'object',
+            additionalProperties: true,
+          },
+        },
+        required: ['name'],
+        additionalProperties: false,
+      },
       case: {
         type: 'object',
         properties: {
           name: {type: 'string'},
+          matrix: {type: 'boolean'},
           description: {type: 'string'},
           requests: {
             type: 'array',
@@ -85,9 +115,25 @@ export default class SuiteLoader {
 
     SuiteLoader.validateSuite(suiteDefinition)
 
+    let cases = suiteDefinition.cases
+
+    if (suiteDefinition.matrix) {
+      const matrixTestCases = cases.filter(({matrix}) => matrix === true)
+      cases = cases.filter(({matrix}) => matrix !== true)
+      suiteDefinition.matrix.forEach(({name, caseName, params, captures, assertions}) => {
+        console.log(`Building matrix test case for ${caseName} as ${name}`)
+        const caseDefinition = matrixTestCases.filter(caseDefinition => caseDefinition.name === caseName).pop()
+        if (!caseDefinition) {
+          throw new Error(`No test case found with name "${caseName}" or it is not annotated as matrix test case`)
+        }
+
+        cases.push(SuiteLoader.prepareMatrixCase(caseDefinition, name, {params, captures, assertions}))
+      })
+    }
+
     return {
       name: suiteDefinition.name,
-      cases: suiteDefinition.cases.map(({name, requests}) =>
+      cases: cases.map(({name, requests}) =>
         SuiteLoader.prepareCase(name, requests, context),
       ),
     }
@@ -103,6 +149,34 @@ export default class SuiteLoader {
         `"${validation.error?.path}" ${validation.error?.message}`,
       )
     }
+  }
+
+  private static prepareMatrixCase({requests}: CaseDefinition, caseName: string, kvBag: KeyValueBag): CaseDefinition {
+    return {name: caseName, requests: requests.map(({request, delayed, params, captures, assertions}) => {
+      params = SuiteLoader.setMatrixValues(params, kvBag.params)
+      captures = SuiteLoader.setMatrixValues(captures, kvBag.captures) as CaptureKeyValue
+      assertions = SuiteLoader.setMatrixValues(assertions, kvBag.assertions)
+
+      return {
+        request,
+        delayed,
+        params,
+        captures,
+        assertions,
+      }
+    })}
+  }
+
+  private static setMatrixValues(caseData: ParamKeyValue|CaptureKeyValue|AssertionKeyValue, matrixData: ParamKeyValue|CaptureKeyValue|AssertionKeyValue): ParamKeyValue|CaptureKeyValue|AssertionKeyValue {
+    Object.entries(caseData || []).forEach(([name, value]) => {
+      if (name === '_') {
+        caseData = matrixData
+      } else if (typeof value === 'string' && value.startsWith('$matrix')) {
+        caseData[name] = matrixData[name]
+      }
+    })
+
+    return caseData
   }
 
   private static prepareCase(