+++ src/helpers/debug.ts
@@ -1,5 +1,5 @@
-export default class Debug {
-  static inspect(val: any): string {
+export const Debug = {
+  inspect: (val: any): string => {
     return require('util').inspect(val, {depth: 10})
-  }
+  },
 }